// src/api/models/portfolioDetailResDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { PortfolioDetailResDTOHopeJob } from './portfolioDetailResDTOHopeJob';

export interface PortfolioDetailResDTO {
  /** @nullable */
  hopeJob: PortfolioDetailResDTOHopeJob;
  id: number;
  name: string;
  description: string;
  responsibilities: string;
  problemSolving: string;
  learnings: string;
  /** @nullable */
  contributionRate: number | null;
  createdAt: string;
}